+++ src/main/java/kitchenpos/application/menu/MenuGroupService.java
@@ -1,0 +1,26 @@
+package kitchenpos.application.menu;
+
+import kitchenpos.domain.menu.MenuGroupRepository;
+import kitchenpos.domain.menu.MenuGroup;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Service
+public class MenuGroupService {
+    private final MenuGroupRepository menuGroupRepository;
+
+    public MenuGroupService(final MenuGroupRepository menuGroupRepository) {
+        this.menuGroupRepository = menuGroupRepository;
+    }
+
+    @Transactional
+    public MenuGroup create(final MenuGroup menuGroup) {
+        return menuGroupRepository.save(menuGroup);
+    }
+
+    public List<MenuGroup> list() {
+        return menuGroupRepository.findAll();
+    }
+}
