+++ src/main/java/kitchenpos/application/menu/MenuService.java
@@ -1,0 +1,84 @@
+package kitchenpos.application.menu;
+
+import kitchenpos.domain.menu.MenuRepository;
+import kitchenpos.domain.menu.MenuGroupRepository;
+import kitchenpos.domain.menu.MenuProductRepository;
+import kitchenpos.domain.product.ProductRepository;
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.menu.MenuProduct;
+import kitchenpos.domain.product.Product;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.math.BigDecimal;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+
+@Service
+public class MenuService {
+    private final MenuRepository menuRepository;
+    private final MenuGroupRepository menuGroupRepository;
+    private final MenuProductRepository menuProductRepository;
+    private final ProductRepository productRepository;
+
+    public MenuService(
+            final MenuRepository menuRepository,
+            final MenuGroupRepository menuGroupRepository,
+            final MenuProductRepository menuProductRepository,
+            final ProductRepository productRepository
+    ) {
+        this.menuRepository = menuRepository;
+        this.menuGroupRepository = menuGroupRepository;
+        this.menuProductRepository = menuProductRepository;
+        this.productRepository = productRepository;
+    }
+
+    @Transactional
+    public Menu create(final Menu menu) {
+        final BigDecimal price = menu.getPrice();
+
+        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
+            throw new IllegalArgumentException();
+        }
+
+        if (!menuGroupRepository.existsById(menu.getMenuGroupId())) {
+            throw new IllegalArgumentException();
+        }
+
+        final List<MenuProduct> menuProducts = menu.getMenuProducts();
+
+        BigDecimal sum = BigDecimal.ZERO;
+        for (final MenuProduct menuProduct : menuProducts) {
+            final Product product = productRepository.findById(menuProduct.getProductId())
+                    .orElseThrow(IllegalArgumentException::new);
+            sum = sum.add(product.getPrice().multiply(BigDecimal.valueOf(menuProduct.getQuantity())));
+        }
+
+        if (price.compareTo(sum) > 0) {
+            throw new IllegalArgumentException();
+        }
+
+        final Menu savedMenu = menuRepository.save(menu);
+
+        final Long menuId = savedMenu.getId();
+        final List<MenuProduct> savedMenuProducts = new ArrayList<>();
+        for (final MenuProduct menuProduct : menuProducts) {
+            menuProduct.setMenuId(menuId);
+            savedMenuProducts.add(menuProductRepository.save(menuProduct));
+        }
+        savedMenu.saveMenuProducts(savedMenuProducts);
+
+        return savedMenu;
+    }
+
+    public List<Menu> list() {
+        final List<Menu> menus = menuRepository.findAll();
+
+        for (final Menu menu : menus) {
+            menu.saveMenuProducts(menuProductRepository.findAllByMenuId(menu.getId()));
+        }
+
+        return menus;
+    }
+}
