+++ src/main/java/kitchenpos/application/order/OrderService.java
@@ -1,0 +1,110 @@
+package kitchenpos.application.order;
+
+import kitchenpos.domain.menu.MenuRepository;
+import kitchenpos.domain.order.OrderDao;
+import kitchenpos.domain.order.OrderLineItemRepository;
+import kitchenpos.domain.order.OrderTableRepository;
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.order.OrderLineItem;
+import kitchenpos.domain.order.OrderStatus;
+import kitchenpos.domain.order.OrderTable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.CollectionUtils;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
+
+@Service
+public class OrderService {
+    private final MenuRepository menuRepository;
+    private final OrderDao orderDao;
+    private final OrderLineItemRepository orderLineItemRepository;
+    private final OrderTableRepository orderTableRepository;
+
+    public OrderService(
+            final MenuRepository menuRepository,
+            final OrderDao orderDao,
+            final OrderLineItemRepository orderLineItemRepository,
+            final OrderTableRepository orderTableRepository
+    ) {
+        this.menuRepository = menuRepository;
+        this.orderDao = orderDao;
+        this.orderLineItemRepository = orderLineItemRepository;
+        this.orderTableRepository = orderTableRepository;
+    }
+
+    @Transactional
+    public Order create(final Order order) {
+        final List<OrderLineItem> orderLineItems = order.getOrderLineItems();
+
+        if (CollectionUtils.isEmpty(orderLineItems)) {
+            throw new IllegalArgumentException();
+        }
+
+        final List<Long> menuIds = orderLineItems.stream()
+                .map(OrderLineItem::getMenuId)
+                .collect(Collectors.toList());
+
+        if (orderLineItems.size() != menuRepository.countByIdIn(menuIds)) {
+            throw new IllegalArgumentException();
+        }
+
+        order.setId(null);
+
+        final OrderTable orderTable = orderTableRepository.findById(order.getOrderTableId())
+                .orElseThrow(IllegalArgumentException::new);
+
+        if (orderTable.isEmpty()) {
+            throw new IllegalArgumentException();
+        }
+
+        order.setOrderTableId(orderTable.getId());
+        order.setOrderStatus(OrderStatus.COOKING.name());
+        order.setOrderedTime(LocalDateTime.now());
+
+        final Order savedOrder = orderDao.save(order);
+
+        final Long orderId = savedOrder.getId();
+        final List<OrderLineItem> savedOrderLineItems = new ArrayList<>();
+        for (final OrderLineItem orderLineItem : orderLineItems) {
+            orderLineItem.setOrderId(orderId);
+            savedOrderLineItems.add(orderLineItemRepository.save(orderLineItem));
+        }
+        savedOrder.setOrderLineItems(savedOrderLineItems);
+
+        return savedOrder;
+    }
+
+    public List<Order> list() {
+        final List<Order> orders = orderDao.findAll();
+
+        for (final Order order : orders) {
+            order.setOrderLineItems(orderLineItemRepository.findAllByOrderId(order.getId()));
+        }
+
+        return orders;
+    }
+
+    @Transactional
+    public Order changeOrderStatus(final Long orderId, final Order order) {
+        final Order savedOrder = orderDao.findById(orderId)
+                .orElseThrow(IllegalArgumentException::new);
+
+        if (Objects.equals(OrderStatus.COMPLETION.name(), savedOrder.getOrderStatus())) {
+            throw new IllegalArgumentException();
+        }
+
+        final OrderStatus orderStatus = OrderStatus.valueOf(order.getOrderStatus());
+        savedOrder.setOrderStatus(orderStatus.name());
+
+        orderDao.save(savedOrder);
+
+        savedOrder.setOrderLineItems(orderLineItemRepository.findAllByOrderId(orderId));
+
+        return savedOrder;
+    }
+}
