+++ src/main/java/kitchenpos/application/order/TableService.java
@@ -1,0 +1,74 @@
+package kitchenpos.application.order;
+
+import kitchenpos.domain.order.OrderDao;
+import kitchenpos.domain.order.OrderTableRepository;
+import kitchenpos.domain.order.OrderStatus;
+import kitchenpos.domain.order.OrderTable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.Objects;
+
+@Service
+public class TableService {
+    private final OrderDao orderDao;
+    private final OrderTableRepository orderTableRepository;
+
+    public TableService(final OrderDao orderDao, final OrderTableRepository orderTableRepository) {
+        this.orderDao = orderDao;
+        this.orderTableRepository = orderTableRepository;
+    }
+
+    @Transactional
+    public OrderTable create(final OrderTable orderTable) {
+        orderTable.setId(null);
+        orderTable.setTableGroupId(null);
+
+        return orderTableRepository.save(orderTable);
+    }
+
+    public List<OrderTable> list() {
+        return orderTableRepository.findAll();
+    }
+
+    @Transactional
+    public OrderTable changeEmpty(final Long orderTableId, final OrderTable orderTable) {
+        final OrderTable savedOrderTable = orderTableRepository.findById(orderTableId)
+                .orElseThrow(IllegalArgumentException::new);
+
+        if (Objects.nonNull(savedOrderTable.getTableGroupId())) {
+            throw new IllegalArgumentException();
+        }
+
+        if (orderDao.existsByOrderTableIdAndOrderStatusIn(
+                orderTableId, Arrays.asList(OrderStatus.COOKING.name(), OrderStatus.MEAL.name()))) {
+            throw new IllegalArgumentException();
+        }
+
+        savedOrderTable.setEmpty(orderTable.isEmpty());
+
+        return orderTableRepository.save(savedOrderTable);
+    }
+
+    @Transactional
+    public OrderTable changeNumberOfGuests(final Long orderTableId, final OrderTable orderTable) {
+        final int numberOfGuests = orderTable.getNumberOfGuests();
+
+        if (numberOfGuests < 0) {
+            throw new IllegalArgumentException();
+        }
+
+        final OrderTable savedOrderTable = orderTableRepository.findById(orderTableId)
+                .orElseThrow(IllegalArgumentException::new);
+
+        if (savedOrderTable.isEmpty()) {
+            throw new IllegalArgumentException();
+        }
+
+        savedOrderTable.setNumberOfGuests(numberOfGuests);
+
+        return orderTableRepository.save(savedOrderTable);
+    }
+}
