+++ src/main/java/kitchenpos/application/product/ProductService.java
@@ -1,0 +1,34 @@
+package kitchenpos.application.product;
+
+import kitchenpos.domain.product.ProductRepository;
+import kitchenpos.domain.product.Product;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.math.BigDecimal;
+import java.util.List;
+import java.util.Objects;
+
+@Service
+public class ProductService {
+    private final ProductRepository productRepository;
+
+    public ProductService(final ProductRepository productRepository) {
+        this.productRepository = productRepository;
+    }
+
+    @Transactional
+    public Product create(final Product product) {
+        final BigDecimal price = product.getPrice();
+
+        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
+            throw new IllegalArgumentException();
+        }
+
+        return productRepository.save(product);
+    }
+
+    public List<Product> list() {
+        return productRepository.findAll();
+    }
+}
