+++ src/main/java/kitchenpos/application/tablegroup/TableGroupService.java
@@ -1,0 +1,89 @@
+package kitchenpos.application.tablegroup;
+
+import kitchenpos.domain.order.OrderDao;
+import kitchenpos.domain.order.OrderTableRepository;
+import kitchenpos.domain.tablegroup.TableGroupRepository;
+import kitchenpos.domain.order.OrderStatus;
+import kitchenpos.domain.order.OrderTable;
+import kitchenpos.domain.tablegroup.TableGroup;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.CollectionUtils;
+
+import java.time.LocalDateTime;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
+
+@Service
+public class TableGroupService {
+    private final OrderDao orderDao;
+    private final OrderTableRepository orderTableRepository;
+    private final TableGroupRepository tableGroupDao;
+
+    public TableGroupService(final OrderDao orderDao, final OrderTableRepository orderTableRepository, final TableGroupRepository tableGroupDao) {
+        this.orderDao = orderDao;
+        this.orderTableRepository = orderTableRepository;
+        this.tableGroupDao = tableGroupDao;
+    }
+
+    @Transactional
+    public TableGroup create(final TableGroup tableGroup) {
+        final List<OrderTable> orderTables = tableGroup.getOrderTables();
+
+        if (CollectionUtils.isEmpty(orderTables) || orderTables.size() < 2) {
+            throw new IllegalArgumentException();
+        }
+
+        final List<Long> orderTableIds = orderTables.stream()
+                .map(OrderTable::getId)
+                .collect(Collectors.toList());
+
+        final List<OrderTable> savedOrderTables = orderTableRepository.findAllByIdIn(orderTableIds);
+
+        if (orderTables.size() != savedOrderTables.size()) {
+            throw new IllegalArgumentException();
+        }
+
+        for (final OrderTable savedOrderTable : savedOrderTables) {
+            if (!savedOrderTable.isEmpty() || Objects.nonNull(savedOrderTable.getTableGroupId())) {
+                throw new IllegalArgumentException();
+            }
+        }
+
+        tableGroup.setCreatedDate(LocalDateTime.now());
+
+        final TableGroup savedTableGroup = tableGroupDao.save(tableGroup);
+
+        final Long tableGroupId = savedTableGroup.getId();
+        for (final OrderTable savedOrderTable : savedOrderTables) {
+            savedOrderTable.setTableGroupId(tableGroupId);
+            savedOrderTable.setEmpty(false);
+            orderTableRepository.save(savedOrderTable);
+        }
+        savedTableGroup.setOrderTables(savedOrderTables);
+
+        return savedTableGroup;
+    }
+
+    @Transactional
+    public void ungroup(final Long tableGroupId) {
+        final List<OrderTable> orderTables = orderTableRepository.findAllByTableGroupId(tableGroupId);
+
+        final List<Long> orderTableIds = orderTables.stream()
+                .map(OrderTable::getId)
+                .collect(Collectors.toList());
+
+        if (orderDao.existsByOrderTableIdInAndOrderStatusIn(
+                orderTableIds, Arrays.asList(OrderStatus.COOKING.name(), OrderStatus.MEAL.name()))) {
+            throw new IllegalArgumentException();
+        }
+
+        for (final OrderTable orderTable : orderTables) {
+            orderTable.setTableGroupId(null);
+            orderTable.setEmpty(false);
+            orderTableRepository.save(orderTable);
+        }
+    }
+}
