+++ src/main/java/kitchenpos/ui/menu/MenuGroupRestController.java
@@ -1,0 +1,37 @@
+package kitchenpos.ui.menu;
+
+import kitchenpos.application.menu.MenuGroupService;
+import kitchenpos.domain.menu.MenuGroup;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+public class MenuGroupRestController {
+    private final MenuGroupService menuGroupService;
+
+    public MenuGroupRestController(final MenuGroupService menuGroupService) {
+        this.menuGroupService = menuGroupService;
+    }
+
+    @PostMapping("/api/menu-groups")
+    public ResponseEntity<MenuGroup> create(@RequestBody final MenuGroup menuGroup) {
+        final MenuGroup created = menuGroupService.create(menuGroup);
+        final URI uri = URI.create("/api/menu-groups/" + created.getId());
+        return ResponseEntity.created(uri)
+                .body(created)
+                ;
+    }
+
+    @GetMapping("/api/menu-groups")
+    public ResponseEntity<List<MenuGroup>> list() {
+        return ResponseEntity.ok()
+                .body(menuGroupService.list())
+                ;
+    }
+}
